Add hmi package and move menu module and action class there
