+++ config/main.cfg
@@ -3,8 +3,9 @@
 FixServer = 127.0.0.1
 FixPort = 3490
 
-[menu]
-config_file = config/menus.yaml
+[hmi]
+# Human Machine Interface Configuration
+config_file = config/hmi.yaml
 
 [mainscreen]
 # Screen Geometry
