+++ pyEfis.py
@@ -39,7 +39,7 @@ except:
 import scheduler
 import fix
 import hooks
-import events
+import hmi
 import gui
 import importlib
 
@@ -74,7 +74,7 @@ if __name__ == "__main__":
         log.info("Starting PyEFIS in %s Mode" % (args.mode,))
 
     log.debug("PyQT Version = %d" % PYQT)
-    events.initialize(config)
+    hmi.initialize(config)
     scheduler.initialize()
 
     host = config.get("main", "FixServer")
